Unroll works with macro_rules macros
The #[unroll_for_loops] attribute has been verified to work inside
macro_rules! macros where the bound is given by a constant expression,
even if that constant is a macro argument. There is nothing new here,
it's just a property of the order of macro expantion in rust. This
commit simply removes the commont saying it may not work in this context.